add validation for endpoints

+++ cars/src/Controller/CarController.php
@@ -16,13 +16,25 @@ use App\Application\Query\GetCarFromId\GetCarFromIdQuery;
 use App\Message\SendEmail;
 use Symfony\Component\Messenger\Stamp\HandledStamp;
 use Exception;
+use App\Domain\Validator\Request\Car\ValidatorCarRequestInterface;
 
 class CarController
 {
+    private $validator;
+
+    public function __construct(ValidatorCarRequestInterface $validator)
+    {
+        $this->validator = $validator;
+    }
+
     public function addCar(Request $request, MessageBusInterface $bus): JsonResponse
     {
         try {
             $data = json_decode($request->getContent(), true);
+
+            if (!$this->validator->validateAddCarRequest($data)) {
+                return new JsonResponse(['error'=> 'bad params'], Response::HTTP_BAD_REQUEST);
+            }
 
             $car = Car::createCar();
             $data['id'] = $car->getId();
@@ -49,6 +61,10 @@ class CarController
         try {
             $data = json_decode($request->getContent(), true);
 
+            if (!$this->validator->validateUpdateCarRequest($data)) {
+                return new JsonResponse(['error'=> 'bad params'], Response::HTTP_BAD_REQUEST);
+            }
+
             $result = $bus->dispatch(new UpdateCarCommand($id, $data))
                 ->last(HandledStamp::class)->getResult();
             $status = ($result['error']) ? Response::HTTP_NOT_FOUND : Response::HTTP_OK;
@@ -69,6 +85,10 @@ class CarController
     public function deleteCar(string $id, MessageBusInterface $bus): JsonResponse
     {
         try {
+            if (!$this->validator->validateDeleteCarRequest($id)) {
+                return new JsonResponse(['error'=> 'bad params'], Response::HTTP_BAD_REQUEST);
+            }
+
             $result = $bus->dispatch(new DeleteCarCommand($id))->last(HandledStamp::class)->getResult();
             $status = ($result['error']) ? Response::HTTP_NOT_FOUND : Response::HTTP_OK;
 
@@ -88,6 +108,10 @@ class CarController
     public function getAllCars(int $page, MessageBusInterface $queryBus): JsonResponse
     {
         try {
+            if (!$this->validator->validateGetAllCarsRequest($page)) {
+                return new JsonResponse(['error'=> 'bad params'], Response::HTTP_BAD_REQUEST);
+            }
+
             $result = $queryBus->dispatch(new GetAllCarsQuery($page))->last(HandledStamp::class)->getResult();
 
             return new JsonResponse($result, Response::HTTP_OK);
@@ -99,6 +123,10 @@ class CarController
     public function getAllCarsEnabled(int $page, MessageBusInterface $queryBus): JsonResponse
     {
         try {
+            if (!$this->validator->validateGetAllCarsEnabledRequest($page)) {
+                return new JsonResponse(['error'=> 'bad params'], Response::HTTP_BAD_REQUEST);
+            }
+
             $result = $queryBus->dispatch(new GetAllCarsEnabledQuery($page))->last(HandledStamp::class)->getResult();
 
             return new JsonResponse($result, Response::HTTP_OK);
@@ -110,6 +138,10 @@ class CarController
     public function getCarFromId(string $id, MessageBusInterface $queryBus): JsonResponse
     {
         try {
+            if (!$this->validator->validateGetCarFromIdRequest($id)) {
+                return new JsonResponse(['error'=> 'bad params'], Response::HTTP_BAD_REQUEST);
+            }
+
             $result = $queryBus->dispatch(new GetCarFromIdQuery($id))->last(HandledStamp::class)->getResult();
 
             $status = ($result['error']) ? Response::HTTP_NOT_FOUND : Response::HTTP_OK;
