+++ cars/src/Controller/UserController.php
@@ -2,6 +2,7 @@
 
 namespace App\Controller;
 
+use App\Domain\Validator\Request\User\ValidatorUserRequestInterface;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -18,10 +19,21 @@ use Exception;
 
 class UserController
 {
+    private $validator;
+
+    public function __construct(ValidatorUserRequestInterface $validator)
+    {
+        $this->validator = $validator;
+    }
+
     public function addUser(Request $request, MessageBusInterface $bus): JsonResponse
     {
         try {
             $data = json_decode($request->getContent(), true);
+
+            if (!$this->validator->validateAddUserRequest($data)) {
+                return new JsonResponse(['error'=> 'bad params'], Response::HTTP_BAD_REQUEST);
+            }
 
             $user = User::createUser();
             $user = $user->buildUserFromArray($user, $data);
@@ -47,6 +59,10 @@ class UserController
         try {
             $data = json_decode($request->getContent(), true);
 
+            if (!$this->validator->validateUpdateUserRequest($data)) {
+                return new JsonResponse(['error'=> 'bad params'], Response::HTTP_BAD_REQUEST);
+            }
+
             $result = $bus->dispatch(new UpdateUserCommand($email, $data))
                 ->last(HandledStamp::class)->getResult();
             $status = ($result['error']) ? Response::HTTP_NOT_FOUND : Response::HTTP_OK;
@@ -67,6 +83,10 @@ class UserController
     public function deleteUser(string $email, MessageBusInterface $bus): JsonResponse
     {
         try {
+            if (!$this->validator->validateDeleteUserRequest($email)) {
+                return new JsonResponse(['error'=> 'bad params'], Response::HTTP_BAD_REQUEST);
+            }
+
             $result = $bus->dispatch(new DeleteUserCommand($email))->last(HandledStamp::class)->getResult();
             $status = ($result['error']) ? Response::HTTP_NOT_FOUND : Response::HTTP_OK;
 
@@ -86,6 +106,10 @@ class UserController
     public function getAllUsers(int $page, MessageBusInterface $queryBus): JsonResponse
     {
         try {
+            if (!$this->validator->validateGetAllUsersRequest($page)) {
+                return new JsonResponse(['error'=> 'bad params'], Response::HTTP_BAD_REQUEST);
+            }
+
             $result = $queryBus->dispatch(new GetAllUsersQuery($page))->last(HandledStamp::class)->getResult();
 
             return new JsonResponse($result, Response::HTTP_OK);
@@ -97,6 +121,10 @@ class UserController
     public function getUserFromEmail(string $email, MessageBusInterface $queryBus): JsonResponse
     {
         try {
+            if (!$this->validator->validateGetUserFromEmailRequest($email)) {
+                return new JsonResponse(['error'=> 'bad params'], Response::HTTP_BAD_REQUEST);
+            }
+
             $result = $queryBus->dispatch(new GetUserFromEmailQuery($email))
                 ->last(HandledStamp::class)->getResult();
